requirements.txt
fastapi==0.116.1
pydantic==2.11.7
uvicorn==0.35.0
requests==2.32.5
psycopg2-binary==2.9.10
httpx==0.28.1
